requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
attrs==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.1.0
openai==1.30.3
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.2
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.11.1
